requirements.txt
amqp==5.0.2
ansible==2.10.4
ansible-base==2.10.3
appdirs==1.4.4
asgiref==3.3.1
asn1crypto==1.4.0
attrs==20.3.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
black==20.8b1
boto==2.49.0
boto3==1.16.30
botocore==1.19.30
celery==5.0.3
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coverage==5.3
cryptography==3.2.1
dj-database-url==0.5.0
Django==3.1.4
django-bootstrap4==2.3.1
django-braces==1.14.0
django-cache-url==3.1.2
django-cors-headers==3.5.0
django-environ==0.4.5
django-filter==2.4.0
django-ipware==3.0.2
django-redis==4.12.1
django-storages==1.10.1
djangorestframework==3.12.2
docopt==0.6.2
docutils==0.16
entrypoints==0.3
flake8==3.8.4
furl==2.1.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.1.1
iso8601==0.1.13
isort==5.6.4
Jinja2==2.11.2
jmespath==0.10.0
jsonfield==3.1.0
kombu==5.0.2
m3u8==0.7.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.6.0
mypy-extensions==0.4.3
mysqlclient==2.0.1
orderedmultidict==1.0.1
packaging==20.7
paramiko==2.7.2
pathspec==0.8.1
prompt-toolkit==3.0.8
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
redis==3.5.3
regex==2020.11.13
requests==2.25.0
s3transfer==0.3.3
sentry-sdk==0.19.4
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
webvtt-py==0.4.6
xmltodict==0.12.0
zipp==3.4.0
